Plenty of new tests and changes
Some temporary code to be removed when I get a better Bank-like "peer"
and when I find a way to get that 97.45% coverage to be nearer 100%.

I also updated the TODO list and ignore list alongside the dependency
list which shouldn't require CircularJSON when I find some solutions in
problems surrounding crypto.js (ie. jrsasign).